Conflicts:
	src/upem/jarret/worker/ClientJarRet.java

+++ src/upem/jarret/worker/Worker.java
@@ -3,7 +3,7 @@ package upem.jarret.worker;
 public interface Worker {
 
 	// Return a JSON String with the result of computing task number taskNumber
-	public String compute(Long taskNumber);
+	public String compute(int taskNumber);
 
 	// Return the jobId of the job treated by this worker
 	public long getJobId();
